requirements.txt
boto3==1.20.31
GeoAlchemy2==0.12.5
geopandas==0.13.0
geopy==2.3.0
ipython==7.33.0
jsonschema==4.17.3
h5py==3.7.0
matplotlib==3.3.4
mercantile==1.2.1
numpy==1.23.3
opencv-python==4.7.0.72
pandas==1.5.1
Pillow==9.5.0
psycopg2==2.9.6
pyproj==3.5.0
rasterio==1.3.6
Requests==2.30.0
scikit_learn==1.0.0
scipy==1.10.1
Shapely==2.0.1
SQLAlchemy==2.0.9
sqlalchemy_utils==0.41.1
tensorflow==2.9.0
tqdm==4.62.3